requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
openai>=1.10.0
qdrant-client>=1.7.0
PyMuPDF>=1.23.8
python-multipart>=0.0.9
sentence-transformers>=2.6.0
torch>=2.0.0
httpx>=0.27.0